Stubbed Game Test Helpers: Separate out StubbedGames classes
