+++ tests/NexusMods.DataModel.Tests/Harness/ADataModelTest.cs
@@ -9,7 +9,7 @@ using NexusMods.Hashing.xxHash64;
 using NexusMods.Paths;
 using NexusMods.Paths.Extensions;
 using NexusMods.Paths.Utilities;
-using NexusMods.StandardGameLocators.TestHelpers;
+using NexusMods.StandardGameLocators.TestHelpers.StubbedGames;
 using Xunit.DependencyInjection;
 // ReSharper disable StaticMemberInGenericType
